+++ lib/rubocop/cop/mixin/heredoc.rb
@@ -4,7 +4,7 @@ module RuboCop
   module Cop
     # Common functionality for working with heredoc strings.
     module Heredoc
-      OPENING_DELIMITER = /<<[~-]?['"`]?([^'"`]+)['"`]?/
+      OPENING_DELIMITER = /(<<[~-]?)['"`]?([^'"`]+)['"`]?/
 
       def on_str(node)
         return unless node.heredoc?
@@ -21,7 +21,11 @@ module RuboCop
       private
 
       def delimiter_string(node)
-        node.source.match(OPENING_DELIMITER).captures.first
+        node.source.match(OPENING_DELIMITER).captures[1]
+      end
+
+      def heredoc_type(node)
+        node.source.match(OPENING_DELIMITER).captures[0]
       end
     end
   end